Add a Page Not Found to show when trying to view a page that does not exist. Added a 'Valid For X hours' message when using the 'generate password-reset' CLI command.
