+++ frontend/src/app/pages/not-found/not-found.component.html
@@ -1,0 +1,9 @@
+<h1 style="margin-top: 3em">The Page Cannot Be Found</h1>
+
+<a mat-flat-button style="margin-top: 3em" routerLink="/">Go Home</a>
+
+<span style="flex-grow: 1"></span>
+
+@if (config?.contactEmail; as contactEmail) {
+  <a style="margin-top: 48px" mat-button [href]="'mailto:' + contactEmail">Contact</a>
+}
