+++ shared/utils.ts
@@ -254,3 +254,49 @@ function sortWildcardParts(aParts: string[], bParts: string[]) {
 
   return
 }
+
+export function humanDuration(ms: number): string {
+  const MINUTE = 60
+  const HOUR = MINUTE * 60
+  const DAY = HOUR * 24
+  const WEEK = DAY * 7
+  const YEAR = DAY * 365.25
+  const MONTH = YEAR / 12
+
+  const seconds = Math.round(ms / 1000)
+  const years = Math.round(seconds / YEAR)
+  const months = Math.round(seconds / MONTH)
+  const weeks = Math.round(seconds / WEEK)
+  const days = Math.round(seconds / DAY)
+  const hours = Math.round(seconds / HOUR)
+  const minutes = Math.round(seconds / MINUTE)
+
+  if (months > 11) {
+    return String(years) + ' year' + ((years > 1) ? 's' : '')
+  }
+
+  if (weeks > 4) {
+    return String(months) + ' month' + ((months > 1) ? 's' : '')
+  }
+
+  if (days > 6) {
+    return String(weeks) + ' week' + ((weeks > 1) ? 's' : '')
+  }
+
+  if (hours > 23) {
+    return String(days) + ' day' + ((days > 1) ? 's' : '')
+  }
+
+  if (minutes > 59) {
+    return String(hours) + ' hour' + ((hours > 1) ? 's' : '')
+  }
+
+  if (seconds > 59) {
+    return String(minutes) + ' minute' + ((minutes > 1) ? 's' : '')
+  }
+
+  if (ms > 999) {
+    return String(seconds) + ' second' + ((seconds > 1) ? 's' : '')
+  }
+  return 'now'
+}
